fixed validatorbehavior

+++ WorkLog.Domain/Entities/User.cs
@@ -5,12 +5,13 @@ namespace WorkLog.Domain.Entities;
 
 public class User: EntityBase
 {
+    #nullable disable
     public string Name { get; set; }
     public string Username { get; set; }
     public string Password { get; set; }
     public string Email { get; set; }
-    public int RoleId { get; set; }
-    public int DepartmentId { get; set; }
+    public long RoleId { get; set; }
+    public long DepartmentId { get; set; }
     public Status Status { get; set; }
     public DateTime LastLoginTime { get; set; }
